+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/DigDurabilityEnchantmentMixin.java
@@ -9,14 +9,14 @@ import org.spongepowered.asm.mixin.injection.At;
 @Mixin(DigDurabilityEnchantment.class)
 public abstract class DigDurabilityEnchantmentMixin {
 	@ModifyReturnValue(at = @At(value = "RETURN"), method = "shouldIgnoreDurabilityDrop")
-	private static boolean onChanceToIgnoreDurabilityDrop(boolean original) {
+	private static boolean iguanatweaksreborn$preventDurabilityDrop(boolean original) {
 		if (EnchantmentsFeature.isUnbreakingOverhaul())
 			return false;
 		return original;
 	}
 
 	@ModifyReturnValue(at = @At(value = "RETURN"), method = "getMaxLevel")
-	private int onChanceToIgnoreDurabilityDrop(int original) {
+	private int iguanatweaksreborn$maxLevel(int original) {
 		if (EnchantmentsFeature.isUnbreakingOverhaul())
 			return 5;
 		return original;