+++ changelog.md
@@ -4,9 +4,10 @@
 ### Changes
 * Increased golden tools efficiency (12 -> 15)
 * Removed flat efficiency bonus and reduced percentage bonus (+0.35 flat bonus -> REMOVED; +25% bonus -> +20% bonus per level)
+* Removed flat unbreaking bonus
 
 ### Fixes
-* Fixed quark integration recipe being wrong
+* Quark integration recipes being in wrong folder
 
 ## 4.13.3
 ### Fixes